Add more sanity checks

+++ models/building_delete.go
@@ -18,5 +18,8 @@ func NewBuildingDelete(pid string) *BuildingDeleteParams {
 
 // Remove delete a row from the store base on id
 func (p *BuildingDeleteParams) Remove(ctx context.Context, store *drivers.Storage) error {
+	if _, oks := store.Exists(p.ID); !oks {
+		return ErrRecordNotFound
+	}
 	return store.Delete(p.ID)
 }
